updates and formatting changes

+++ src/com/zombies/Listeners/OnEntityDamageEvent.java
@@ -53,7 +53,8 @@ public class OnEntityDamageEvent implements Listener
 						Entity entity = damager.getDamager();
 						if (!(plugin.manager.isEntityInGame(entity)))
 						{
-							if (plugin.manager.isPlayerInGame((Player) e.getEntity()))
+							if (plugin.manager.isPlayerInGame((Player) e
+									.getEntity()))
 							{
 								e.setCancelled(true);
 							}
@@ -62,10 +63,12 @@ public class OnEntityDamageEvent implements Listener
 						{
 							final Player player = (Player) e.getEntity();
 							Game game = plugin.manager.getGame(player);
-							int damage = 6;
-							if (game.perkManager.getPlayersPerks().containsKey(player))
+							double damage = 6;
+							if (game.perkManager.getPlayersPerks().containsKey(
+									player))
 							{
-								if (game.perkManager.getPlayersPerks().get(player).contains(PerkType.JUGGERNOG))
+								if (game.perkManager.getPlayersPerks().get(
+										player).contains(PerkType.JUGGERNOG))
 								{
 									damage = damage / 2;
 								}
@@ -78,21 +81,22 @@ public class OnEntityDamageEvent implements Listener
 							}
 							else
 							{
-								if(game.downedPlayerManager.isPlayerDowned(player))
+								if (game.downedPlayerManager
+										.isPlayerDowned(player))
 								{
 									e.setCancelled(true);
 								}
 								e.setDamage(damage);
 							}
-							Bukkit.getScheduler().scheduleSyncDelayedTask(plugin, new Runnable()
-							{
-								
-								public void run()
-								{
-									healPlayer(player);
-								}
-								
-							}, 100L);
+							Bukkit.getScheduler().scheduleSyncDelayedTask(
+									plugin, new Runnable() {
+										
+										public void run()
+										{
+											healPlayer(player);
+										}
+										
+									}, 100L);
 						}
 					}
 				}
@@ -101,8 +105,9 @@ public class OnEntityDamageEvent implements Listener
 		else if (e.getEntity() instanceof Zombie)
 		{
 			Entity entity = e.getEntity();
-			int damage = 0;
-			if (!(plugin.manager.isEntityInGame(entity))) return;
+			double damage = 0;
+			if (!(plugin.manager.isEntityInGame(entity)))
+				return;
 			Game game = plugin.manager.getGame(entity);
 			if (game != null)
 			{
@@ -110,18 +115,23 @@ public class OnEntityDamageEvent implements Listener
 				if (e.getDamager() instanceof Player)
 				{
 					Player player = (Player) e.getDamager();
-					if (player.getItemInHand().getType().equals(Material.IRON_SWORD))
+					if (player.getItemInHand().getType().equals(
+							Material.IRON_SWORD))
 					{
 						if (game.players.contains(player))
 						{
 							Zombie zombie1 = (Zombie) entity;
 							double damageAmount = e.getDamage();
-							int totalHealth;
-							double cx = player.getLocation().getX() - zombie1.getLocation().getX();
-							double cy = player.getLocation().getY() - zombie1.getLocation().getY();
-							double cz = player.getLocation().getZ() - zombie1.getLocation().getZ();
+							Double totalHealth;
+							double cx = player.getLocation().getX()
+									- zombie1.getLocation().getX();
+							double cy = player.getLocation().getY()
+									- zombie1.getLocation().getY();
+							double cz = player.getLocation().getZ()
+									- zombie1.getLocation().getZ();
 							
-							if (Math.sqrt(Math.pow(cx, 2) + Math.pow(cy, 2) + Math.pow(cz, 2)) <= plugin.config.meleeRange)
+							if (Math.sqrt(Math.pow(cx, 2) + Math.pow(cy, 2)
+									+ Math.pow(cz, 2)) <= plugin.config.meleeRange)
 							{
 								damageAmount = 5;
 							}
@@ -130,51 +140,63 @@ public class OnEntityDamageEvent implements Listener
 								e.setCancelled(true);
 								return;
 							}
-							if (game.spawnManager.totalHealth().containsKey(e.getEntity()))
+							if (game.spawnManager.totalHealth().containsKey(
+									e.getEntity()))
 							{
-								totalHealth = game.spawnManager.totalHealth().get(e.getEntity());
+								totalHealth = game.spawnManager.totalHealth()
+										.get(e.getEntity());
 							}
 							else
 							{
 								game.spawnManager.setTotalHealth(entity, 20);
-								totalHealth = 20;
+								totalHealth = 20D;
 							}
 							if (totalHealth >= 20)
 							{
-								zombie1.setHealth(20);
+								zombie1.setHealth(20D);
 								if (game.isDoublePoints())
 								{
-									plugin.pointManager.addPoints(player, plugin.config.pointsOnHit * 2);
+									plugin.pointManager.addPoints(player,
+											plugin.config.pointsOnHit * 2);
 								}
 								else
 								{
-									plugin.pointManager.addPoints(player, plugin.config.pointsOnHit);
+									plugin.pointManager.addPoints(player,
+											plugin.config.pointsOnHit);
 								}
-								game.spawnManager.setTotalHealth(e.getEntity(), (int) (totalHealth - damageAmount));
-								if (game.spawnManager.totalHealth().get(e.getEntity()) < 20)
+								game.spawnManager.setTotalHealth(e.getEntity(),
+										(int) (totalHealth - damageAmount));
+								if (game.spawnManager.totalHealth().get(
+										e.getEntity()) < 20)
 								{
-									zombie1.setHealth(game.spawnManager.totalHealth().get(e.getEntity()));
+									zombie1.setHealth(game.spawnManager
+											.totalHealth().get(e.getEntity()));
 								}
 								plugin.pointManager.notifyPlayer(player);
 							}
-							else if (totalHealth < 1 || totalHealth - damageAmount <= 1)
+							else if (totalHealth < 1
+									|| totalHealth - damageAmount <= 1)
 							{
 								e.setCancelled(true);
-								OnZombiePerkDrop perkdrop = new OnZombiePerkDrop(plugin);
+								OnZombiePerkDrop perkdrop = new OnZombiePerkDrop(
+										plugin);
 								perkdrop.perkDrop(zombie1, player);
 								zombie1.remove();
 								boolean doublePoints = game.isDoublePoints();
 								if (doublePoints)
 								{
-									plugin.pointManager.addPoints(player, plugin.config.pointsOnKill * 2);
+									plugin.pointManager.addPoints(player,
+											plugin.config.pointsOnKill * 2);
 								}
 								else
 								{
-									plugin.pointManager.addPoints(player, plugin.config.pointsOnKill);
+									plugin.pointManager.addPoints(player,
+											plugin.config.pointsOnKill);
 								}
 								zombie1.playEffect(EntityEffect.DEATH);
 								plugin.pointManager.notifyPlayer(player);
-								game.spawnManager.removeEntity((Entity) zombie1);
+								game.spawnManager
+										.removeEntity((Entity) zombie1);
 								game.zombieKilled(player);
 							}
 							else
@@ -182,23 +204,28 @@ public class OnEntityDamageEvent implements Listener
 								zombie1.damage(damage);
 								if (game.isDoublePoints())
 								{
-									plugin.pointManager.addPoints(player, plugin.config.pointsOnHit * 2);
+									plugin.pointManager.addPoints(player,
+											plugin.config.pointsOnHit * 2);
 								}
 								else
 								{
-									plugin.pointManager.addPoints(player, plugin.config.pointsOnHit);
+									plugin.pointManager.addPoints(player,
+											plugin.config.pointsOnHit);
 								}
 								plugin.pointManager.notifyPlayer(player);
 							}
-							game.spawnManager.setTotalHealth(e.getEntity(), (int) (totalHealth - damageAmount));
+							game.spawnManager.setTotalHealth(e.getEntity(),
+									(int) (totalHealth - damageAmount));
 							if (game.isInstaKill())
 							{
 								zombie1.remove();
-								game.spawnManager.removeEntity((Entity) zombie1);
+								game.spawnManager
+										.removeEntity((Entity) zombie1);
 							}
 							for (Player pl : game.players)
 							{
-								pl.playSound(entity.getLocation().add(0, 1, 0), Sound.STEP_STONE, 1, 1);
+								pl.playSound(entity.getLocation().add(0, 1, 0),
+										Sound.STEP_STONE, 1, 1);
 							}
 						}
 						else
@@ -222,17 +249,20 @@ public class OnEntityDamageEvent implements Listener
 		if (e.getEntity() instanceof Player)
 		{
 			Player player = (Player) e.getEntity();
-			if( plugin.manager.getGame(player) == null)
+			if (plugin.manager.getGame(player) == null)
 				return;
-			if ( plugin.manager.getGame(player).downedPlayerManager.isPlayerDowned(player))
+			if (plugin.manager.getGame(player).downedPlayerManager
+					.isPlayerDowned(player))
 			{
 				e.setCancelled(true);
 			}
-			if (plugin.manager.getGame(player) != null && plugin.manager.getGame(player).mode == ArenaStatus.STARTING)
+			if (plugin.manager.getGame(player) != null
+					&& plugin.manager.getGame(player).mode == ArenaStatus.STARTING)
 			{
 				e.setCancelled(true);
 			}
-			if (player.getHealth() < 1 || player.getHealth() - e.getDamage() < 1)
+			if (player.getHealth() < 1
+					|| player.getHealth() - e.getDamage() < 1)
 			{
 				if (plugin.manager.isPlayerInGame(player))
 				{
@@ -244,13 +274,17 @@ public class OnEntityDamageEvent implements Listener
 					}
 				}
 			}
-			if (plugin.manager.isPlayerInGame(player)) player.getLocation().getWorld().playEffect(player.getLocation().add(0, 1, 0), Effect.STEP_SOUND, 152);
+			if (plugin.manager.isPlayerInGame(player))
+				player.getLocation().getWorld().playEffect(
+						player.getLocation().add(0, 1, 0), Effect.STEP_SOUND,
+						152);
 		}
-		else if(e.getCause().equals(DamageCause.LAVA) && e.getEntity() instanceof Zombie)
+		else if (e.getCause().equals(DamageCause.LAVA)
+				&& e.getEntity() instanceof Zombie)
 		{
 			Zombie z = (Zombie) e.getEntity();
-			Game game =  plugin.manager.getGame(z);
-			if(game == null)
+			Game game = plugin.manager.getGame(z);
+			if (game == null)
 				return;
 			z.setFireTicks(0);
 			z.teleport(game.getPlayerSpawn());
@@ -266,15 +300,19 @@ public class OnEntityDamageEvent implements Listener
 		}
 		if (!game.downedPlayerManager.isPlayerDowned(player))
 		{
-			Bukkit.broadcastMessage(COMZombies.prefix + player.getName() + " Has gone down! Stand close and right click him to revive");
+			Bukkit.broadcastMessage(COMZombies.prefix
+					+ player.getName()
+					+ " Has gone down! Stand close and right click him to revive");
 			DownedPlayer down = new DownedPlayer(player, game);
 			down.setPlayerDown(true);
 			game.downedPlayerManager.addDownedPlayer(down);
-			player.setHealth(1);
+			player.setHealth(1D);
 		}
-		if(game.downedPlayerManager.getDownedPlayers().size() == game.players.size())
+		if (game.downedPlayerManager.getDownedPlayers().size() == game.players
+				.size())
 		{
-			for (DownedPlayer downedPlayer : game.downedPlayerManager.getDownedPlayers())
+			for (DownedPlayer downedPlayer : game.downedPlayerManager
+					.getDownedPlayers())
 			{
 				downedPlayer.cancelDowned();
 			}
@@ -285,11 +323,13 @@ public class OnEntityDamageEvent implements Listener
 	
 	public void healPlayer(final Player player)
 	{
-		if (beingHealed.contains(player)) return;
-		else beingHealed.add(player);
-		if (!(plugin.manager.isPlayerInGame(player))) return;
-		Bukkit.getScheduler().scheduleSyncDelayedTask(plugin, new Runnable()
-		{
+		if (beingHealed.contains(player))
+			return;
+		else
+			beingHealed.add(player);
+		if (!(plugin.manager.isPlayerInGame(player)))
+			return;
+		Bukkit.getScheduler().scheduleSyncDelayedTask(plugin, new Runnable() {
 			
 			@Override
 			public void run()
@@ -311,12 +351,14 @@ public class OnEntityDamageEvent implements Listener
 	
 	public void removeDownedPlayer(Player player)
 	{
-		plugin.manager.getGame(player).downedPlayerManager.removeDownedPlayer(player);
+		plugin.manager.getGame(player).downedPlayerManager
+				.removeDownedPlayer(player);
 	}
 	
 	@SuppressWarnings("deprecation")
 	public boolean isDownedPlayer(String name)
 	{
-		return plugin.manager.getGame(Bukkit.getPlayer(name)).downedPlayerManager.isPlayerDowned(Bukkit.getPlayer(name));
+		return plugin.manager.getGame(Bukkit.getPlayer(name)).downedPlayerManager
+				.isPlayerDowned(Bukkit.getPlayer(name));
 	}
 }
