+++ src/com/zombies/config/ConfigSetup.java
@@ -124,7 +124,7 @@ public class ConfigSetup
 		reviveRange = plugin.getConfig().getInt("config.ReviveSettings.ReviveRange");
 		if (reviveRange > 6)
 			reviveRange = 6;
-		meleeRange = plugin.getConfig().getFloat("config.ReviveSettings.MeleeRange");
+		meleeRange = (float) plugin.getConfig().getDouble("config.ReviveSettings.MeleeRange");
 		configVersion = plugin.getConfig().getString("vID");
 		reloadTime = plugin.getConfig().getInt("config.gameSettings.reloadTime");
 		if (plugin.possibleGuns.size() != 0)
